requirements.txt
flwr==1.22.0
torch==2.2.2
torchvision==0.17.2
numpy==1.26.4
scikit-learn==1.1.3
pillow==10.2.0
facenet-pytorch==2.6.0
